Moved update functionality to it’s own module
